requirements.txt
joblib==1.1.0
matplotlib==3.3.4
numpy==1.19.5
opencv_python_headless==4.3.0.36
pybullet==3.2.5
PyYAML==6.0
scipy==1.5.4
torch==1.7.0
torchgeometry==0.1.2
tqdm==4.64.0
treelib==1.6.1
